+++ src/Process/ProcessSet.php
@@ -207,17 +207,17 @@ class ProcessSet implements \Countable
      * Set queued processes without delay as prepared
      * @param OutputInterface $output If provided, list of dequeued and queued processes will be printed
      */
-    public function dequeueProcessesWithoutDelay(OutputInterface $output = null)
+    public function dequeueProcessesWithoutDelay(OutputInterface $output)
     {
         $queuedProcesses = $this->get(self::PROCESS_STATUS_QUEUED);
         foreach ($queuedProcesses as $className => $processObject) {
             if ($processObject->delayMinutes === null) {
-                if ($output) {
+                if ($output->isDebug()) {
                     $output->writeln(sprintf('Testcase "%s" is prepared to be run', $className));
                 }
                 $this->setStatus($className, self::PROCESS_STATUS_PREPARED);
             } else {
-                if ($output) {
+                if ($output->isDebug()) {
                     $output->writeln(
                         sprintf(
                             'Testcase "%s" is queued to be run %01.1f minutes after testcase "%s" is finished',