+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 ### Added
 - Provide information about results of finished processes (verbose (`-v`) mode of `run` command must be enabled)
 
+### Changed
+- Adjust output verbosity for different verbosity levels to improve clarity of the output mainly when running the tests locally.
+    - Use default level if you care only about test results count.
+    - Use verbose mode (`-v`) to see name of failed tests during execution.
+    - Use very verbose mode (`-vv`) to see detailed progress information during execution and also output of failed tests.
+    - Debug level (`-vvv`) should be used when you want to know all available information about the run or when you run the tests on CI server. The tests output is printed incrementally.
+
 ## 1.2.0 - 2016-01-11
 ### Added
 - Property annotation added to `SyntaxSugarTrait` to know `$wd` property meaning in IDE. ([#36](https://github.com/lmc-eu/steward/pull/36))